plugins/requirements.txt
disco-py==0.0.11